# api/openapi.yaml
openapi: 3.1.0
info:
  title: BotHub API
  version: 0.1.0
  description: |
    Contract between the Go dashboard and the PHP API.

    Rules:
    - The API is the only way to change configuration. The dashboard never touches the DB.
    - Every route except `/api/health`, `GET /api/v1/setup`, `POST /api/v1/setup` and
      `POST /api/v1/auth/login` needs a valid session. One central middleware checks this.
    - Every changing request (POST, PUT, PATCH, DELETE) needs the header `X-CSRF-Token`.
    - Errors never contain UI text. They contain an i18n key that the dashboard translates.
    - Bot tokens are write-only. The API never returns them.
    - Actions that need Discord return `202` with a job. Poll `GET /api/v1/jobs/{jobId}`.
servers:
  - url: /
security:
  - session: []
...
paths:
...
  /api/v1/bots/{botId}/commands/deleted/{commandId}/restore:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - $ref: "#/components/parameters/CommandId"
    post:
      operationId: restoreDeletedCommand
      tags: [commands]
      summary: Bring a deleted command back
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Command
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/bots/{botId}/command-groups:
    parameters:
      - $ref: "#/components/parameters/BotId"
    get:
      operationId: listCommandGroups
      tags: [commands]
      summary: Groups (folders) of the custom commands, by position
      responses:
        "200":
          description: Groups
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items: { $ref: "#/components/schemas/CommandGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      operationId: createCommandGroup
      tags: [commands]
      summary: Create a group (max. 50 per bot)
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommandGroupWrite" }
      responses:
        "201":
          description: Group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/bots/{botId}/command-groups/{groupId}:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - name: groupId
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    put:
      operationId: updateCommandGroup
      tags: [commands]
      summary: Rename or move a group
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CommandGroupWrite" }
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CommandGroup" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationError" }
    delete:
      operationId: deleteCommandGroup
      tags: [commands]
      summary: Delete a group; its commands become ungrouped
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "204": { description: Deleted }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/bots/{botId}/commands/{commandId}/versions/{versionId}/restore:
    parameters:
      - $ref: "#/components/parameters/BotId"
      - $ref: "#/components/parameters/CommandId"
      - name: versionId
        in: path
        required: true
        schema: { type: integer, minimum: 1 }
    post:
      operationId: restoreCommandVersion
      tags: [commands]
      summary: Make a saved version the current graph (stored as a new version)
      parameters:
        - $ref: "#/components/parameters/CsrfToken"
      responses:
        "200":
          description: Command
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Command" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: bothub_session
...
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      description: Token from `GET /api/v1/auth/me`
      schema: { type: string }
    BotId:
      name: botId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
    CommandId:
      name: commandId
      in: path
      required: true
      schema: { type: integer, minimum: 1 }
...
  responses:
    Unauthorized:
      description: No valid session (`error.auth.required`, `error.auth.invalid_credentials`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: CSRF token missing or wrong (`error.csrf.invalid`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Resource not found (`error.<resource>.not_found`)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: State does not allow this action
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    ValidationError:
      description: Request body invalid
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [key]
          properties:
            key:
              type: string
              description: i18n key, e.g. `error.bot.not_found`
              examples: [error.bot.not_found]
            params:
              type: object
              description: Values for placeholders in the translated text
              additionalProperties: true
            fields:
              type: object
              description: Per-field i18n keys for validation errors
              additionalProperties: { type: string }
...
    CommandGroup:
      type: object
      required: [id, name, position, commands]
      properties:
        id: { type: integer }
        name: { type: string, maxLength: 40 }
        description: { type: string, maxLength: 200 }
        position: { type: integer, minimum: 0, maximum: 999, description: Lower numbers show first }
        commands: { type: integer, description: Number of commands in the group }
    CommandGroupWrite:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 40 }
        description: { type: string, maxLength: 200 }
        position: { type: integer, minimum: 0, maximum: 999 }
...
    ModuleKey:
      type: string
      pattern: "^[a-z][a-z0-9-]{1,63}$"
      examples: [economy, leveling, ticket]
...
    CommandSummary:
      type: object
      required: [id, name, builtin, modified, enabled]
      properties:
        id: { type: integer }
        name:
          type: string
          pattern: "^[a-z0-9_-]{1,32}( [a-z0-9_-]{1,32}){0,2}$"
        module:
          description: Owning module for built-in commands
          oneOf: [{ $ref: "#/components/schemas/ModuleKey" }, { type: "null" }]
        builtin: { type: boolean }
        modified:
          type: boolean
          description: Built-in command whose graph differs from the default
        enabled: { type: boolean }
        groupId:
          description: Command group (folder), or null
          oneOf: [{ type: integer }, { type: "null" }]
        updatedAt: { type: string, format: date-time }
    Command:
      allOf:
        - $ref: "#/components/schemas/CommandSummary"
        - type: object
          required: [graph]
          properties:
            graph:
              type: object
              description: See shared/graph.schema.json
...